abstracting feedback component and container
